Remove unnecessary cast
git-svn-id: http://ehcache-spring-annotations.googlecode.com/svn/trunk@390 e9fffd1a-4083-750a-f54f-a748c1f33b81

+++ src/main/java/com/googlecode/ehcache/annotations/config/EhCacheConfigBeanDefinitionParser.java
@@ -71,7 +71,7 @@ BeanDefinitionParser {
 			final Element evictExpiredElement = (Element)evictExpiredElements.item(0);
 
 			final int interval = Integer.parseInt(evictExpiredElement.getAttribute(XSD_ATTRIBUTE__INTERVAL));
-			final long longInterval = MSEC_PER_MINUTE * ((long) interval);
+			final long longInterval = MSEC_PER_MINUTE * interval;
 
 			List<CacheNameMatcher> cacheNameMatchers = parseEvictExpiredElement(evictExpiredElement);
